requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.1
django-coverage-plugin==1.3.1
Sphinx
django-extensions==1.6.7
Werkzeug==0.11.10
django-test-plus==1.0.13
factory_boy==2.7.0
sqlparse<0.2
django-debug-toolbar==1.4

# improved REPL
ipdb==0.10.0

pytest-django==2.9.1
pytest-sugar==0.7.1
